handlers/mentor_handlers.py
from telegram import Update
from telegram.ext import ContextTypes
import logging

logger = logging.getLogger(__name__)

async def handle_mentor(update: Update, context: ContextTypes.DEFAULT_TYPE):
    try:
        # Здесь будет логика работы наставников с приглашенными
        await update.message.reply_text(
            "Система работы наставников:\n\n"
            "• Просмотр прогресса команды\n"
            "• Консультации и поддержка\n"
            "• Проверка выполнения заданий\n"
            "• Назначение встреч и вебинаров\n\n"
            "Выберите действие для продолжения."
        )
    except Exception as e:
        logger.error(f"Ошибка в handle_mentor: {e}")
        await update.message.reply_text("Произошла ошибка при обработке запроса.")

async def handle_team_progress(update: Update, context: ContextTypes.DEFAULT_TYPE):
    try:
        # Здесь будет логика показа прогресса команды
        await update.message.reply_text(
            "Прогресс вашей команды:\n\n"
            "• Всего участников: 10\n"
            "• Активных: 8\n"
            "• Выполнили план на неделю: 6\n"
            "• Средний процент выполнения: 85%"
        )
    except Exception as e:
        logger.error(f"Ошибка в handle_team_progress: {e}")
        await update.message.reply_text("Произошла ошибка при обработке запроса.")