requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gspread==5.11.3
oauth2client==4.1.3
pandas==2.0.3